+++ powertrader-ts/backend/src/engine/backtest/IBacktestEngine.js
@@ -1,4 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const IStrategy_1 = require("../strategy/IStrategy");
-//# sourceMappingURL=IBacktestEngine.js.map